pyproject.toml
[project]
name = "brain-mri-analysis-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=8.29.0",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "pyqt6>=6.7.1",
    "scikit-image>=0.24.0",
    "tk>=0.1.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
